use a modfule for a node (#46)
